ESCKAN-97 - reuse ingestion command for new composer igestoin service
